requirements.txt
bcrypt==3.2.0
cffi==1.15.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
mathparse==0.1.2
mysql-connector==2.2.9
nltk==3.7
packaging==21.3
Pint==0.18
pycparser==2.21
pymongo==3.12.3
pyparsing==3.0.7
python-dateutil==2.7.5
PyYAML==3.13
regex==2022.3.15
six==1.16.0
SQLAlchemy==1.2.19
tqdm==4.63.0
Werkzeug==2.0.2
WTForms==3.0.1
